OOFB-56 Adicionando conversão de bigint para int

+++ ferramentas-auxiliares/retirada-de-informaçoes-sobre-os-consentimentos/attachments/consent_function_consolidated_stock.sql
@@ -5,8 +5,8 @@ CREATE OR REPLACE FUNCTION consent_consolidated_stock()
 )
 LANGUAGE SQL
 AS $$
-select  SUM(CASE WHEN tb.sha_business_document_number IS NULL THEN 1 ELSE 0 END) AS Clientes_Unicos_PF_Total,
-		SUM(CASE WHEN tb.sha_business_document_number IS NOT NULL THEN 1 ELSE 0 END) AS Clientes_Unicos_PJ_Total
+select  SUM(CASE WHEN tb.sha_business_document_number IS NULL THEN 1 ELSE 0 END)::int AS Clientes_Unicos_PF_Total,
+		SUM(CASE WHEN tb.sha_business_document_number IS NOT NULL THEN 1 ELSE 0 END)::int AS Clientes_Unicos_PJ_Total
 from (
 	SELECT distinct c.sha_business_document_number,
 	c.sha_person_document_number
